pyproject.toml
[tool.poetry]
name = "lawyer-assist"
version = "0.1.0"
description = ""
authors = ["tcstrength <thai.cuong1404@gmail.com>"]
readme = "README.md"
packages = [
    {include = "src/api"},
    {include = "src/data"},
    {include = "src/gui"},
    {include = "src/util"}
]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
ipykernel = "^6.29.5"
lxml = "^5.2.2"
html2text = "^2024.2.26"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
tqdm = "^4.66.4"
cohere = "^5.5.8"
setuptools = "^70.3.0"
transformers = "^4.42.4"
mlx = "^0.16.0"
sentencepiece = "^0.2.0"
protobuf = "^5.27.2"
pyyaml = "^6.0.1"
jinja2 = "^3.1.4"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
sentence-transformers = "^3.0.1"
mlx-lm = "^0.15.3"
pyvi = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
